refactor: consume shared artifact contracts
