+++ pyproject.toml
@@ -69,6 +69,7 @@ requires-python = ">=3.12"
 
 # Core dependencies
 dependencies = [
+    "ml4t-data>=0.1.0b7",
     "polars>=0.20.0",
     "pandas>=2.0.0",
     "numpy>=1.24.0",
@@ -101,6 +102,7 @@ comparison = [
 
 # Development dependencies
 dev = [
+    "ml4t-diagnostic>=0.1.0b1",
     "pytest>=7.4.0",
     "pytest-cov>=4.1.0",
     "pytest-xdist>=3.3.0",
@@ -135,6 +137,7 @@ Changelog = "https://github.com/ml4t/backtest/blob/main/CHANGELOG.md"
 
 [dependency-groups]
 dev = [
+    "ml4t-diagnostic>=0.1.0b1",
     "pytest>=7.4.0",
     "pytest-cov>=4.1.0",
     "pytest-xdist>=3.3.0",
@@ -248,5 +251,7 @@ exclude_lines = [
 ]
 
 [tool.uv.sources]
+ml4t-data = { path = "../ml4t-data", editable = true }
+ml4t-engineer = { path = "../ml4t-engineer", editable = true }
 # Sibling package for optional tearsheet integration
 ml4t-diagnostic = { path = "../ml4t-diagnostic", editable = true }
