+++ src/ml4t/backtest/_validation_imports.py
@@ -12,7 +12,6 @@ from .execution.impact import LinearImpact
 from .execution.limits import VolumeParticipationLimit
 from .execution.rebalancer import RebalanceConfig, TargetWeightExecutor
 from .execution.schedule import RebalanceCadence, RebalanceSchedule, resolve_rebalance_timestamps
-from .feed_spec import FeedSpec
 from .models import (
     FixedSlippage,
     NoCommission,
@@ -39,7 +38,6 @@ __all__ = [
     "BacktestConfig",
     "DataFeed",
     "Engine",
-    "FeedSpec",
     "ExecutionMode",
     "Strategy",
     "Order",
