+++ tests/contracts/test_public_api_surface.py
@@ -13,7 +13,6 @@ def test_root_api_contains_only_intended_core_surface() -> None:
         "BacktestConfig",
         "BacktestResult",
         "CommissionType",
-        "FeedSpec",
         "OrderType",
         "OrderSide",
         "OrderStatus",
