+++ tests/test_config_wiring.py
@@ -12,12 +12,12 @@ Validates that BacktestConfig fields actually affect execution:
 from datetime import datetime
 
 import pytest
+from ml4t.data.artifacts.market_data import FeedSpec
 
 from ml4t.backtest import (
     BacktestConfig,
     Broker,
     ExecutionMode,
-    FeedSpec,
 )
 from ml4t.backtest.config import (
     CommissionType,
