+++ tests/test_equity_curve.py
@@ -3,8 +3,9 @@
 from datetime import datetime, timedelta
 
 import polars as pl
+from ml4t.data.artifacts.market_data import FeedSpec
 
-from ml4t.backtest import BacktestConfig, DataFeed, Engine, FeedSpec, Strategy
+from ml4t.backtest import BacktestConfig, DataFeed, Engine, Strategy
 from ml4t.backtest.analytics.equity import EquityCurve
 from ml4t.backtest.config import DataFrequency
